Integratp pru-pwmcss per il controllo motori

+++ src/events/MyOutMotors.cpp
@@ -8,7 +8,7 @@
 #include <commons/MyPriority.h>
 #include <events/MyOutMotors.h>
 
-MyOutMotors::MyOutMotors(boost::uuids::uuid origin, int32_t front, int32_t rear, int32_t left, int32_t right) : MyEvent(origin), front(front), rear(rear), left(left), right(right) {
+MyOutMotors::MyOutMotors(boost::uuids::uuid origin, uint16_t front, uint16_t rear, uint16_t left, uint16_t right) : MyEvent(origin), front(front), rear(rear), left(left), right(right) {
 	this->setPriority(MyPriority::OUT_MOTORS_PRIORITY);
 
 }
@@ -20,15 +20,15 @@ MyEvent::EventType MyOutMotors::getType() const {
 	return MyEvent::EventType::OutMotors;
 }
 
-int32_t MyOutMotors::getFront() const {
+uint16_t MyOutMotors::getFront() const {
 	return front;
 }
-int32_t MyOutMotors::getRear() const {
+uint16_t MyOutMotors::getRear() const {
 	return rear;
 }
-int32_t MyOutMotors::getLeft() const {
+uint16_t MyOutMotors::getLeft() const {
 	return left;
 }
-int32_t MyOutMotors::getRight() const {
+uint16_t MyOutMotors::getRight() const {
 	return right;
 }
